+++ docs/changelog.rst
@@ -14,6 +14,7 @@ New Features
 - Add the SCons target definition equivalent to the ECMF and CMake-simulation cartesian product parameterized simulation
   files (:issue:`61`, :merge:`64`). By `Kyle Brindley`_.
 - Record the Abaqus environment for each Abaqus builder task (:issue:`85`, :merge:`75`). By `Kyle Brindley`_.
+- Add prototype Cubit geometry tutorial source files (:issue:`180`, :merge:`76`). By `Kyle Brindley`_.
 
 Documentation
 =============